pyproject.toml
[project]
name = "rssgen"
version = "0.1.0"
description = "自托管 RSS 源生成框架"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "beautifulsoup4>=4.13",
    "feedgen>=1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "cachetools>=5.5",
]
